pyproject.toml
[project]
name = "customer-support"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.1",
    "langchain-google-genai>=3.0.0",
    "langchain-tavily>=0.2.12",
    "langgraph>=1.0.1",
    "python-dotenv>=1.1.1",
]
